Move nidx_et_cache to avoid weird imports
